div/requirements.txt
Flask==2.0.2
requests==2.22.0
opentelemetry-instrumentation-flask==0.27b0
opentelemetry-instrumentation-requests==0.27b0
gunicorn